requirements.txt
Django==1.7
psycopg2==2.5.4
gunicorn==19.1.1
django-ckeditor==4.4.6
